deploy: version 1.5

+++ src/pages/api/routes/programs/[id].ts
@@ -2,6 +2,9 @@ import type { NextApiRequest, NextApiResponse } from "next";
 
 import { prisma } from "../../../../auth/db";
 
+import { getServerSession } from "next-auth";
+import { authOptions } from 'auth/auth';
+
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   const id = req.query.id as string;
 
@@ -13,9 +16,21 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
         },
       });
 
+      let user = null;
+
+      // from field program.createdBy , get user
+      if (program?.createdBy) {
+        user = await prisma.user.findUnique({
+          where: {
+            id: program.createdBy,
+          },
+        });
+      }
+
+
       return res
         .status(200)
-        .json({ message: `Program fetched.`, data: program });
+        .json({ message: `Program fetched.`, data: { ...program, creator: user?.name } });
     } catch (error) {
       return res
         .status(500)
@@ -25,6 +40,11 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   }
 
   if (req.method === "DELETE") {
+    const session = await getServerSession(req, res, authOptions);
+    if (!session) {
+      res.status(401).json({ message: `Unauthorized.` });
+      return;
+    }
     try {
       const program = await prisma.onboardingProgram.delete({
         where: {
@@ -44,6 +64,11 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   }
 
   if (req.method === "PATCH") {
+    const session = await getServerSession(req, res, authOptions);
+    if (!session) {
+      res.status(401).json({ message: `Unauthorized.` });
+      return;
+    }
     try {
       const { name, content } = req.body as {
         name: string;
